chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-remove-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
FAIL hello.import should be non-null. Was undefined
Node removal should make import deleted.
FAIL hello.import should be null (of type object). Was undefined (of type undefined).
Re-inserting it shouldn't help.
FAIL hello.import should be non-null. Was undefined
PASS successfullyParsed is true

TEST COMPLETE
